+++ packages/server-services/sql/views/completed_module_item_count_view.sql
@@ -2,7 +2,7 @@ SELECT
     cicv.user_id,
     cicv.course_id,
     cicv.module_id,
-    COUNT(*) completed_course_item_count
+    COUNT(*)::int completed_course_item_count
 FROM public.course_item_completion_view cicv
 
 GROUP BY cicv.user_id, cicv.course_id, cicv.module_id
